Separate into Issuer and Verifier classes.
