+++ lib/Issuer.js
@@ -1,18 +1,14 @@
 /*!
  * Copyright (c) 2022 Digital Bazaar, Inc. All rights reserved.
  */
-const {httpClient} = require('@digitalbazaar/http-client');
-const {ZcapClient} = require('@digitalbazaar/ezcap');
-const https = require('https');
-const didKey = require('@digitalbazaar/did-method-key');
-const {decodeSecretKeySeed} = require('bnid');
-const {Ed25519Signature2020} = require('@digitalbazaar/ed25519-signature-2020');
+import {httpClient} from '@digitalbazaar/http-client/main.js';
+import {ZcapClient} from '@digitalbazaar/ezcap';
+import {decodeSecretKeySeed} from 'bnid';
+import {Ed25519Signature2020} from '@digitalbazaar/ed25519-signature-2020';
 import {constructOAuthHeader} from './oath2';
+import {agent, didKeyDriver} from './constants.js';
 
-const agent = new https.Agent({rejectUnauthorized: false});
-const didKeyDriver = didKey.driver();
-
-class Issuer {
+export class Issuer {
   constructor({oath2, issuer}) {
     this.oath2 = oath2;
     this.issuer = issuer;
@@ -39,6 +35,29 @@ class Issuer {
       json: body
     });
   }
+  async setStatus({body}) {
+    const headers = {
+      ...this.settings.issuer.headers
+    };
+    try {
+      let result;
+      if(this.settings.issuer.zcap) {
+        result = await _zcapClientRequest({
+          endpoint: this.settings.issuer.statusEndpoint,
+          zcap: this.settings.zcap,
+          json: body
+        });
+      } else {
+        result = await httpClient.post(
+          this.settings.issuer.statusEndpoint,
+          {headers, agent, json: body});
+      }
+      return result;
+    } catch(e) {
+      console.log(e);
+      throw e;
+    }
+  }
 }
 
 async function _httpRequest({endpoint, json, headers = {}, oath2}) {
@@ -58,7 +77,7 @@ async function _httpRequest({endpoint, json, headers = {}, oath2}) {
   } catch(e) {
     // delete the Authorization header to prevent
     // oath2 headers in reports
-    e?.response?.headers.delete('Authorization');
+    e.response.headers.delete('Authorization');
     error = e;
   }
   return {result, error};
@@ -98,5 +117,3 @@ async function _zcapClientRequest({endpoint, zcap, json}) {
   }
   return {result, error};
 }
-
-module.exports = {Issuer};
